feat: add importing, add support for multiprocessing
Goal here is to try and log data while the feeder is going. We will stick with a set rate for now. Later will figure out a way. Hopefully the files are readable. Currently using a tmp file to mark that the feeder is running to end the logging loop. logging every few seconds